Version 2
Resolved a bug that resulted in loss of image resolution during the resizing process. Previously, the image was downscaled, resized to the Region of Interest (ROI), and then resized again to the target size, leading to a reduction in image quality. The fix ensures that the image is resized directly to the ROI and then immediately to the final resolution, preserving higher image quality.

Additionally, included boxplots to visually represent the size distribution of the ROIs across different classes.